requirements.txt
anyio==3.6.1
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
firebase-admin==6.0.0
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
hyperframe==6.0.1
idna==3.4
msgpack==1.0.4
oauthlib==3.2.1
proto-plus==1.22.1
protobuf==4.21.7
pyapns-client==2.0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
